feat: move language code validation to transform
