+++ application/CohortManager/compose.core.yaml
@@ -300,7 +300,6 @@ services:
       - ExceptionFunctionURL=http://create-exception:7070/api/CreateException
       - BsSelectGpPracticeUrl=http://bs-select-gp-practice-data-service:7998/api/BsSelectGpPracticeDataService
       - BsSelectOutCodeUrl=http://bs-select-outcode-data-service:7881/api/BsSelectOutCodeDataService
-      - LanguageCodeUrl=http://language-code-data-service:7997/api/LanguageCodeDataService
       - CurrentPostingUrl=http://current-posting-data-service:7996/api/CurrentPostingDataService
       - ExcludedSMULookupUrl=http://excluded-smu-data-service:7995/api/ExcludedSMUDataService
       - BlobContainerName=rules
