+++ application/CohortManager/src/Functions/CohortDistributionServices/TransformDataService/Program.cs
@@ -15,6 +15,7 @@ var host = hostBuilder.ConfigureFunctionsWorkerDefaults()
     .AddDataServicesHandler()
         .AddDataServiceStaticCachedClient<BsSelectOutCode>(config.BsSelectOutCodeUrl)
         .AddDataServiceStaticCachedClient<BsSelectGpPractice>(config.BsSelectGpPracticeUrl)
+        .AddDataServiceStaticCachedClient<LanguageCode>(config.LanguageCodeUrl)
         .AddDataService<CohortDistribution>(config.CohortDistributionDataServiceUrl)
         .Build()
     .ConfigureServices(services =>
