+++ application/CohortManager/src/Functions/CohortDistributionServices/TransformDataService/TransformDataService.cs
@@ -28,12 +28,15 @@ public class TransformDataService
     private readonly IExceptionHandler _exceptionHandler;
     private readonly ITransformReasonForRemoval _transformReasonForRemoval;
     private readonly IDataServiceClient<CohortDistribution> _cohortDistributionClient;
+    private readonly ITransformDataLookupFacade _dataLookup;
+
     public TransformDataService(
         ICreateResponse createResponse,
         IExceptionHandler exceptionHandler,
         ILogger<TransformDataService> logger,
         ITransformReasonForRemoval transformReasonForRemoval,
-        IDataServiceClient<CohortDistribution> cohortDistributionClient
+        IDataServiceClient<CohortDistribution> cohortDistributionClient,
+        ITransformDataLookupFacade dataLookup
     )
     {
         _createResponse = createResponse;
@@ -41,6 +44,7 @@ public class TransformDataService
         _logger = logger;
         _transformReasonForRemoval = transformReasonForRemoval;
         _cohortDistributionClient = cohortDistributionClient;
+        _dataLookup = dataLookup;
     }
 
     [Function("TransformDataService")]
@@ -135,6 +139,7 @@ public class TransformDataService
         var ruleParameters = new[] {
             new RuleParameter("databaseParticipant", databaseParticipant),
             new RuleParameter("participant", participant),
+            new RuleParameter("dbLookup", _dataLookup)
         };
 
         var resultList = await re.ExecuteAllRulesAsync("TransformData", ruleParameters);
