+++ application/CohortManager/src/Functions/CohortDistributionServices/TransformDataService/TransformDataServiceConfig.cs
@@ -1,9 +1,17 @@
 namespace NHS.CohortManager.CohortDistribution;
 
+using System.ComponentModel.DataAnnotations;
+
 public class TransformDataServiceConfig
 {
-    public string ExceptionFunctionURL {get;set;}
-    public string BsSelectOutCodeUrl {get;set;}
-    public string BsSelectGpPracticeUrl {get;set;}
-    public string CohortDistributionDataServiceUrl {get;set;}
+    [Required]
+    public string ExceptionFunctionURL { get; set; }
+    [Required]
+    public string BsSelectOutCodeUrl { get; set; }
+    [Required]
+    public string BsSelectGpPracticeUrl { get; set; }
+    [Required]
+    public string CohortDistributionDataServiceUrl { get; set; }
+    [Required]
+    public string LanguageCodeUrl { get; set; }
 }
