+++ application/CohortManager/src/Functions/ScreeningValidationService/LookupValidation/Breast_Screening_lookupRules.json
@@ -67,24 +67,6 @@
         }
       },
       {
-        "RuleName": "00.ValidateLanguageCode.NBO.NonFatal",
-        "LocalParams": [
-          {
-            "Name": "languageCode",
-            "Expression": "newParticipant.PreferredLanguage"
-          }
-        ],
-        "Expression": "string.IsNullOrEmpty(languageCode) || dbLookup.ValidateLanguageCode(languageCode)",
-        "Actions": {
-          "OnFailure": {
-            "Name": "OutputExpression",
-            "Context": {
-              "Expression": "\"Invalid preferred language code\""
-            }
-          }
-        }
-      },
-      {
         "RuleName": "11.ValidateReasonForRemoval.NBO.NonFatal",
         "LocalParams": [
           {
