+++ application/CohortManager/src/Functions/ScreeningValidationService/LookupValidation/IDataLookupFacadeBreastScreening.cs
@@ -3,7 +3,6 @@ public interface IDataLookupFacadeBreastScreening
 {
     bool CheckIfPrimaryCareProviderExists(string primaryCareProvider);
     bool ValidateOutcode(string postcode);
-    bool ValidateLanguageCode(string languageCode);
     bool CheckIfCurrentPostingExists(string currentPosting);
     bool ValidatePostingCategories(string currentPosting);
     bool CheckIfPrimaryCareProviderInExcludedSmuList(string primaryCareProvider);
