+++ application/CohortManager/src/Functions/ScreeningValidationService/LookupValidation/LookupValidationConfig.cs
@@ -4,7 +4,6 @@ public class LookupValidationConfig
     public string ExceptionFunctionUrl {get;set;}
     public string BsSelectGpPracticeUrl {get;set;}
     public string BsSelectOutCodeUrl {get;set;}
-    public string LanguageCodeUrl {get;set;}
     public string CurrentPostingUrl {get;set;}
     public string ExcludedSMULookupUrl {get;set;}
     public string BlobContainerName {get;set;}
