+++ application/CohortManager/src/Functions/ScreeningValidationService/LookupValidation/Program.cs
@@ -16,7 +16,6 @@ var host = hostBuilder.ConfigureFunctionsWorkerDefaults()
     .AddDataServicesHandler()
         .AddDataServiceStaticCachedClient<BsSelectGpPractice>(config.BsSelectGpPracticeUrl)
         .AddDataServiceStaticCachedClient<BsSelectOutCode>(config.BsSelectOutCodeUrl)
-        .AddDataServiceStaticCachedClient<LanguageCode>(config.LanguageCodeUrl)
         .AddDataServiceStaticCachedClient<CurrentPosting>(config.CurrentPostingUrl)
         .AddDataServiceStaticCachedClient<ExcludedSMULookup>(config.ExcludedSMULookupUrl)
         .Build()
